pyproject.toml
[tool.poetry]
name = "final-project"
version = "0.1.0"
description = "Django warehouse management system"
authors = ["Your Name <email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2"
djangorestframework = "^3.14"
djangorestframework-simplejwt = "^5.3"
mysqlclient = "^2.2"

[tool.poetry.group.dev.dependencies]
black = "^23.0"
flake8 = "^6.0"
pytest = "^7.4"
pytest-django = "^4.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py"]